a little dashboard refactoring
- moved ManageYourVAHealthCare component to its own file to prepare for additional work
- moved some functions to the helpers file
