+++ src/applications/personalization/dashboard/components/ManageYourVAHealthCare.jsx
@@ -1,0 +1,62 @@
+import React from 'react';
+
+import {
+  DowntimeNotification,
+  externalServices,
+} from 'platform/monitoring/DowntimeNotification';
+
+import {
+  recordDashboardClick,
+  renderWidgetDowntimeNotification,
+} from '../helpers';
+
+import MessagingWidget from '../containers/MessagingWidget';
+import PrescriptionsWidget from '../containers/PrescriptionsWidget';
+
+const ScheduleAnAppointmentWidget = () => (
+  <div id="rx-widget">
+    <h3>Schedule an appointment</h3>
+    <p>
+      Find out how to make a doctor’s appointment with a member of your VA
+      health care team online or by phone.
+    </p>
+    <p>
+      <a
+        href="/health-care/schedule-view-va-appointments/"
+        onClick={recordDashboardClick('schedule-appointment')}
+      >
+        Schedule an appointment
+      </a>
+    </p>
+  </div>
+);
+
+const ManageYourVAHealthCare = () => (
+  <>
+    <h2>Manage your VA health care</h2>
+    <DowntimeNotification
+      appTitle="messaging"
+      dependencies={[externalServices.mvi, externalServices.mhv]}
+      render={renderWidgetDowntimeNotification(
+        'Secure messaging',
+        'Track Secure Messages',
+      )}
+    >
+      <MessagingWidget />
+    </DowntimeNotification>
+
+    <DowntimeNotification
+      appTitle="rx"
+      dependencies={[externalServices.mvi, externalServices.mhv]}
+      render={renderWidgetDowntimeNotification(
+        'prescription refill',
+        'Refill Prescriptions',
+      )}
+    >
+      <PrescriptionsWidget />
+    </DowntimeNotification>
+    <ScheduleAnAppointmentWidget />
+  </>
+);
+
+export default ManageYourVAHealthCare;
