+++ src/applications/personalization/dashboard/containers/DashboardAppNew.jsx
@@ -3,6 +3,8 @@ import { connect } from 'react-redux';
 import Scroll from 'react-scroll';
 import { withRouter } from 'react-router';
 
+import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
+
 import backendServices from 'platform/user/profile/constants/backendServices';
 import { selectProfile } from 'platform/user/selectors';
 import recordEvent from 'platform/monitoring/record-event';
@@ -10,17 +12,12 @@ import localStorage from 'platform/utilities/storage/localStorage';
 
 import { removeSavedForm as removeSavedFormAction } from '../actions';
 
-import FormList from '../components/FormList';
-import MessagingWidget from './MessagingWidget';
-import ClaimsAppealsWidget from './ClaimsAppealsWidget';
-import PrescriptionsWidget from './PrescriptionsWidget';
-import PreferencesWidget from 'applications/personalization/preferences/containers/PreferencesWidget';
+import { recordDashboardClick } from '../helpers';
 
-import {
-  DowntimeNotification,
-  externalServices,
-} from 'platform/monitoring/DowntimeNotification';
-import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
+import FormList from '../components/FormList';
+import ManageYourVAHealthCare from '../components/ManageYourVAHealthCare';
+import ClaimsAppealsWidget from './ClaimsAppealsWidget';
+import PreferencesWidget from 'applications/personalization/preferences/containers/PreferencesWidget';
 
 import profileManifest from 'applications/personalization/profile360/manifest.json';
 import accountManifest from 'applications/personalization/account/manifest.json';
@@ -35,49 +32,6 @@ const scrollToTop = () => {
     smooth: true,
   });
 };
-
-const renderWidgetDowntimeNotification = (appName, sectionTitle) => (
-  downtime,
-  children,
-) => {
-  switch (downtime.status) {
-    case 'down':
-      return (
-        <div>
-          <h2>{sectionTitle}</h2>
-          <AlertBox
-            content={
-              <div>
-                <h4 className="usa-alert-heading">
-                  {appName} is down for maintenance
-                </h4>
-                <p>
-                  We’re making some updates to our {appName.toLowerCase()} tool.
-                  We’re sorry it’s not working right now and hope to be finished
-                  by {downtime.startTime.format('MMMM Do')},{' '}
-                  {downtime.endTime.format('LT')}. Please check back soon.
-                </p>
-              </div>
-            }
-            isVisible
-            status="warning"
-          />
-        </div>
-      );
-    default:
-      return children;
-  }
-};
-
-function recordDashboardClick(product, actionType = 'view-link') {
-  return () => {
-    recordEvent({
-      event: 'dashboard-navigation',
-      'dashboard-action': actionType,
-      'dashboard-product': product,
-    });
-  };
-}
 
 const EmptyStateLinks = () => (
   <div>
@@ -134,24 +88,6 @@ const EmptyStateLinks = () => (
   </div>
 );
 
-const ScheduleAnAppointmentWidget = () => (
-  <div id="rx-widget">
-    <h3>Schedule an appointment</h3>
-    <p>
-      Find out how to make a doctor’s appointment with a member of your VA
-      health care team online or by phone.
-    </p>
-    <p>
-      <a
-        href="/health-care/schedule-view-va-appointments/"
-        onClick={recordDashboardClick('schedule-appointment')}
-      >
-        Schedule an appointment
-      </a>
-    </p>
-  </div>
-);
-
 const ManageBenefitsOrRequestRecords = () => (
   <>
     <h2>Manage benefits or request records</h2>
@@ -193,34 +129,6 @@ const ManageBenefitsOrRequestRecords = () => (
         </a>
       </li>
     </ul>
-  </>
-);
-
-const ManageYourVAHealthCare = () => (
-  <>
-    <h2>Manage your VA health care</h2>
-    <DowntimeNotification
-      appTitle="messaging"
-      dependencies={[externalServices.mvi, externalServices.mhv]}
-      render={renderWidgetDowntimeNotification(
-        'Secure messaging',
-        'Track Secure Messages',
-      )}
-    >
-      <MessagingWidget />
-    </DowntimeNotification>
-
-    <DowntimeNotification
-      appTitle="rx"
-      dependencies={[externalServices.mvi, externalServices.mhv]}
-      render={renderWidgetDowntimeNotification(
-        'prescription refill',
-        'Refill Prescriptions',
-      )}
-    >
-      <PrescriptionsWidget />
-    </DowntimeNotification>
-    <ScheduleAnAppointmentWidget />
   </>
 );
 
@@ -349,13 +257,7 @@ class DashboardAppNew extends React.Component {
             <p>
               <a
                 href={facilityLocator.rootUrl}
-                onClick={() => {
-                  recordEvent({
-                    event: 'dashboard-navigation',
-                    'dashboard-action': 'view-link',
-                    'dashboard-product': 'find-center',
-                  });
-                }}
+                onClick={recordDashboardClick('find-center')}
               >
                 Find your nearest VA medical center
               </a>
@@ -381,6 +283,7 @@ class DashboardAppNew extends React.Component {
       canAccessAppeals,
       profile,
       removeSavedForm,
+      showManageYourVAHealthCare,
     } = this.props;
     const availableWidgetsCount = [
       canAccessClaims,
@@ -416,7 +319,7 @@ class DashboardAppNew extends React.Component {
 
         {availableWidgetsCount === 0 && <EmptyStateLinks />}
 
-        <ManageYourVAHealthCare />
+        {showManageYourVAHealthCare && <ManageYourVAHealthCare />}
         <ManageBenefitsOrRequestRecords />
         <ViewYourProfile />
         <ManageYourAccount />
@@ -454,6 +357,7 @@ const mapStateToProps = state => {
     canAccessAppeals,
     canAccessClaims,
     profile: profileState,
+    showManageYourVAHealthCare: canAccessRx || canAccessMessaging,
   };
 };
