+++ src/applications/personalization/dashboard/helpers.jsx
@@ -1,4 +1,10 @@
+import React from 'react';
 import Raven from 'raven-js';
+
+import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
+
+import recordEvent from 'platform/monitoring/record-event';
+
 import hcaManifest from '../../hca/manifest.js';
 import dependentStatusManifest from '../../disability-benefits/686/manifest.js';
 import feedbackManifest from '../../edu-benefits/feedback-tool/manifest.js';
@@ -169,3 +175,47 @@ export const isFormAuthorizable = formConfig => !!formConfig.authorize;
 
 export const getFormAuthorizationState = (formConfig, state) =>
   formConfig.getAuthorizationState(state);
+
+export const recordDashboardClick = (
+  product,
+  actionType = 'view-link',
+) => () => {
+  recordEvent({
+    event: 'dashboard-navigation',
+    'dashboard-action': actionType,
+    'dashboard-product': product,
+  });
+};
+
+export const renderWidgetDowntimeNotification = (appName, sectionTitle) => (
+  downtime,
+  children,
+) => {
+  switch (downtime.status) {
+    case 'down':
+      return (
+        <div>
+          <h2>{sectionTitle}</h2>
+          <AlertBox
+            content={
+              <div>
+                <h4 className="usa-alert-heading">
+                  {appName} is down for maintenance
+                </h4>
+                <p>
+                  We’re making some updates to our {appName.toLowerCase()} tool.
+                  We’re sorry it’s not working right now and hope to be finished
+                  by {downtime.startTime.format('MMMM Do')},{' '}
+                  {downtime.endTime.format('LT')}. Please check back soon.
+                </p>
+              </div>
+            }
+            isVisible
+            status="warning"
+          />
+        </div>
+      );
+    default:
+      return children;
+  }
+};
